Refactor Doppler renderer and document refactor tasks
