+++ src/heart/display/renderers/doppler/__init__.py
@@ -1,0 +1,5 @@
+from heart.display.renderers.doppler.provider import \
+    DopplerStateProvider as DopplerStateProvider
+from heart.display.renderers.doppler.renderer import \
+    DopplerRenderer as DopplerRenderer
+from heart.display.renderers.doppler.state import DopplerState as DopplerState
